fix: Refactor About Us page
