--- src/components/aboutUs/AboutUsView.tsx
@@ -1,55 +1,0 @@
-import React from "react";
-
-// Link to route
-import { Link } from "react-router-dom";
-
-//MUI styles
-import Stack from "@mui/material/Stack";
-import Typography from "@mui/material/Typography";
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
-import Divider from "@mui/material/Divider";
-
-//import child components files
-import StorySection from "./StorySection";
-import TeamCards from "./TeamCards";
-
-const AboutUsView = () => {
-    return (
-        <Stack spacing={5}>
-            <StorySection />
-            <Divider />
-            <Typography sx={styles.teamTitle}>Meet the Team</Typography>
-            <TeamCards />
-            <Divider />
-            <Box sx={styles.buttonBox}>
-                <Button
-                    component={Link}
-                    to="/contact"
-                    variant="contained"
-                    color="success"
-                    size="small"
-                >
-                    Contact Us
-                </Button>
-            </Box>
-            <Divider />
-        </Stack>
-    );
-};
-
-export default AboutUsView;
-
-const styles = {
-    teamTitle: {
-        fontSize: "30px",
-        fontFamily: "Cursive ",
-        textAlign: "center",
-    },
-
-    buttonBox: {
-        width: "100%",
-        display: "flex",
-        justifyContent: "center",
-    },
-};
