+++ src/pages/__tests__/AboutUs.test.tsx
@@ -3,9 +3,13 @@ import { shallow } from "enzyme";
 import AboutUs from "../AboutUs";
 
 // test component rendering
-describe("Test AboutUs component", () => {
+describe("Test AboutUs page", () => {
     const wrapper = shallow(<AboutUs />);
-    it("Should be able to render AboutUsView component ", () => {
-        expect(wrapper.find("AboutUsView").children).toHaveLength(1);
+    it("Should be able to render StorySection component ", () => {
+        expect(wrapper.find("StorySection").children).toHaveLength(1);
+    });
+
+    it("Should be able to render TeamCards component ", () => {
+        expect(wrapper.find("TeamCards").children).toHaveLength(1);
     });
 });
